MAINT: use gdtrix, gdtria, nrdtrimn, nrdtrisd from xsf
