+++ scipy/special/_add_newdocs.py
@@ -3876,63 +3876,6 @@ add_newdoc("gdtrc",
     True
     """)
 
-add_newdoc("gdtria",
-    """
-    gdtria(p, b, x, out=None)
-
-    Inverse of `gdtr` vs a.
-
-    Returns the inverse with respect to the parameter `a` of ``p =
-    gdtr(a, b, x)``, the cumulative distribution function of the gamma
-    distribution.
-
-    Parameters
-    ----------
-    p : array_like
-        Probability values.
-    b : array_like
-        `b` parameter values of `gdtr(a, b, x)`. `b` is the "shape" parameter
-        of the gamma distribution.
-    x : array_like
-        Nonnegative real values, from the domain of the gamma distribution.
-    out : ndarray, optional
-        If a fourth argument is given, it must be a numpy.ndarray whose size
-        matches the broadcast result of `a`, `b` and `x`.  `out` is then the
-        array returned by the function.
-
-    Returns
-    -------
-    a : scalar or ndarray
-        Values of the `a` parameter such that ``p = gdtr(a, b, x)`.  ``1/a``
-        is the "scale" parameter of the gamma distribution.
-
-    See Also
-    --------
-    gdtr : CDF of the gamma distribution.
-    gdtrib : Inverse with respect to `b` of `gdtr(a, b, x)`.
-    gdtrix : Inverse with respect to `x` of `gdtr(a, b, x)`.
-    gammaincinv : Inverse of the incomplete regularized gamma function.
-
-    Notes
-    -----
-    `gdtria` is implemented in terms of the incomplete gamma inverse as
-    ``gdtria(p, b, x) = gammaincinv(b, p)/x``.
-
-    Examples
-    --------
-    First evaluate `gdtr`.
-
-    >>> from scipy.special import gdtr, gdtria
-    >>> p = gdtr(1.2, 3.4, 5.6)
-    >>> print(p)
-    0.94378087442
-
-    Verify the inverse.
-
-    >>> gdtria(p, 3.4, 5.6)
-    1.2
-    """)
-
 add_newdoc("gdtrib",
     """
     gdtrib(a, p, x, out=None)
@@ -4008,66 +3951,6 @@ add_newdoc("gdtrib",
     >>> gdtrib(1.2, p, 5.6)
     3.3999999999999995
     """)
-
-add_newdoc("gdtrix",
-    """
-    gdtrix(a, b, p, out=None)
-
-    Inverse of `gdtr` vs x.
-
-    Returns the inverse with respect to the parameter `x` of ``p =
-    gdtr(a, b, x)``, the cumulative distribution function of the gamma
-    distribution. This is also known as the pth quantile of the
-    distribution.
-
-    Parameters
-    ----------
-    a : array_like
-        `a` parameter values of ``gdtr(a, b, x)``. ``1/a`` is the "scale"
-        parameter of the gamma distribution.
-    b : array_like
-        `b` parameter values of ``gdtr(a, b, x)``. `b` is the "shape" parameter
-        of the gamma distribution.
-    p : array_like
-        Probability values.
-    out : ndarray, optional
-        If a fourth argument is given, it must be a numpy.ndarray whose size
-        matches the broadcast result of `a`, `b` and `x`. `out` is then the
-        array returned by the function.
-
-    Returns
-    -------
-    x : scalar or ndarray
-        Values of the `x` parameter such that `p = gdtr(a, b, x)`.
-
-    See Also
-    --------
-    gdtr : CDF of the gamma distribution.
-    gdtria : Inverse with respect to `a` of ``gdtr(a, b, x)``.
-    gdtrib : Inverse with respect to `b` of ``gdtr(a, b, x)``.
-    gammaincinv : Inverse of the incomplete regularized gamma function.
-
-    Notes
-    -----
-    `gdtrix` is implemented in terms of the incomplete gamma inverse as
-    ``gdtrix(a, b, p) = gammaincinv(b, p)/a``.
-
-    Examples
-    --------
-    First evaluate `gdtr`.
-
-    >>> from scipy.special import gdtr, gdtrix
-    >>> p = gdtr(1.2, 3.4, 5.6)
-    >>> print(p)
-    0.94378087442
-
-    Verify the inverse.
-
-    >>> gdtrix(1.2, 3.4, p)
-    5.6
-    """)
-
-
 
 add_newdoc("huber",
     r"""
@@ -5959,113 +5842,6 @@ add_newdoc("nctdtrit",
 
     >>> nctdtrit(3, 1, p)
     array([0.5, 1. , 1.5])
-
-    """)
-
-add_newdoc("nrdtrimn",
-    """
-    nrdtrimn(p, std, x, out=None)
-
-    Calculate mean of normal distribution given other params.
-
-    Parameters
-    ----------
-    p : array_like
-        CDF values, in range (0, 1].
-    std : array_like
-        Standard deviation.
-    x : array_like
-        Quantiles, i.e. the upper limit of integration.
-    out : ndarray, optional
-        Optional output array for the function results
-
-    Returns
-    -------
-    mn : scalar or ndarray
-        The mean of the normal distribution.
-
-    See Also
-    --------
-    scipy.stats.norm : Normal distribution
-    ndtr : Standard normal cumulative probability distribution
-    ndtri : Inverse of standard normal CDF with respect to quantile
-    nrdtrisd : Inverse of normal distribution CDF with respect to
-               standard deviation
-
-    Examples
-    --------
-    `nrdtrimn` can be used to recover the mean of a normal distribution
-    if we know the CDF value `p` for a given quantile `x` and the
-    standard deviation `std`. First, we calculate
-    the normal distribution CDF for an exemplary parameter set.
-
-    >>> from scipy.stats import norm
-    >>> mean = 3.
-    >>> std = 2.
-    >>> x = 6.
-    >>> p = norm.cdf(x, loc=mean, scale=std)
-    >>> p
-    0.9331927987311419
-
-    Verify that `nrdtrimn` returns the original value for `mean`.
-
-    >>> from scipy.special import nrdtrimn
-    >>> nrdtrimn(p, std, x)
-    3.0000000000000004
-
-    """)
-
-add_newdoc("nrdtrisd",
-    """
-    nrdtrisd(mn, p, x, out=None)
-
-    Calculate standard deviation of normal distribution given other params.
-
-    Parameters
-    ----------
-    mn : scalar or ndarray
-        The mean of the normal distribution.
-    p : array_like
-        CDF values, in range (0, 1].
-    x : array_like
-        Quantiles, i.e. the upper limit of integration.
-
-    out : ndarray, optional
-        Optional output array for the function results
-
-    Returns
-    -------
-    std : scalar or ndarray
-        Standard deviation.
-
-    See Also
-    --------
-    scipy.stats.norm : Normal distribution
-    ndtr : Standard normal cumulative probability distribution
-    ndtri : Inverse of standard normal CDF with respect to quantile
-    nrdtrimn : Inverse of normal distribution CDF with respect to
-               mean
-
-    Examples
-    --------
-    `nrdtrisd` can be used to recover the standard deviation of a normal
-    distribution if we know the CDF value `p` for a given quantile `x` and
-    the mean `mn`. First, we calculate the normal distribution CDF for an
-    exemplary parameter set.
-
-    >>> from scipy.stats import norm
-    >>> mean = 3.
-    >>> std = 2.
-    >>> x = 6.
-    >>> p = norm.cdf(x, loc=mean, scale=std)
-    >>> p
-    0.9331927987311419
-
-    Verify that `nrdtrisd` returns the original value for `std`.
-
-    >>> from scipy.special import nrdtrisd
-    >>> nrdtrisd(mean, p, x)
-    2.0000000000000004
 
     """)
