+++ scipy/special/xsf_wrappers.h
@@ -338,9 +338,9 @@ double xsf_chdtrc(double df, double x);
 double xsf_chdtri(double df, double y);
 double xsf_gdtr(double a, double b, double x);
 double xsf_gdtrc(double a, double b, double x);
-double special_gdtria(double p, double b, double x);
-double special_gdtrix(double a, double b, double p);
+double xsf_gdtria(double p, double b, double x);
 double xsf_gdtrib(double a, double p, double x);
+double xsf_gdtrix(double a, double b, double p);
 double xsf_kolmogorov(double x);
 double xsf_kolmogc(double x);
 double xsf_kolmogi(double x);
@@ -354,8 +354,8 @@ npy_cdouble xsf_cndtr(npy_cdouble x);
 double xsf_log_ndtr(double x);
 npy_cdouble xsf_clog_ndtr(npy_cdouble x);
 double xsf_ndtri(double x);
-double special_nrdtrimn(double p, double std, double x);
-double special_nrdtrisd(double mean, double p, double x);
+double xsf_nrdtrimn(double p, double std, double x);
+double xsf_nrdtrisd(double mean, double p, double x);
 double xsf_owens_t(double h, double a);
 double xsf_pdtr(double k, double m);
 double xsf_pdtrc(double k, double m);
